AIS / PSS - Corrections for v6 (#182)
The following corrections should be done before finalizing AIS and PSS v6:

**AIS**

- The example for the the property _links->self in the
AccountTransactionReport schema is wrong
/accounts/550e8400e29b11d4a716446655440000/statements ->
/accounts/550e8400e29b11d4a716446655440000/transactions
- GET /iso20022/statements & GET /iso20022/reports:
The current operation ID "listIso20022AccountStatementIds" suggest, that
a list of IDs is returned. This is not entirely true, rather a list of
references is returned. I suggest to rename the operation ID to
"listIso20022AccountStatementReferences"
- GET /iso20022/statements & GET /iso20022/reports:
The summary and description claim that a list of resource links is
returned. The response does not contain any links.

**PSS**

- typo on line 853: stauts -> status
- Fix camel case of schema name StandingorderSubmissionRequest ->
StandingOrderSubmissionRequest

**Both**

- X-Next-Cursor Header should be changed to:
  - required =true
  - schema-type = nullable

---------

Co-authored-by: sfti bot <github-bot@common-api.ch>
Co-authored-by: Sven Biellmann <121558620+svenbiellmann@users.noreply.github.com>